refactor: Introduced an interface for CursorService and moved to common module
