+++ components/common/src/main/java/com/bloxbean/cardano/yaci/store/common/domain/HealthStatus.java
@@ -1,4 +1,4 @@
-package com.bloxbean.cardano.yaci.store.core.domain;
+package com.bloxbean.cardano.yaci.store.common.domain;
 
 import lombok.Builder;
 import lombok.Data;
